Added cooldown for broadcasted messages
A cooldown timer was added into each class file for the commands so that
when a player runs a command, it logs the system time that they ran the
command and logs it into the config.yml. After a certain amount of time,
when the player runs the command again, it will re-broadcast the
message.

This has not been tested yet. It is recommended to wait for a BukkitDev
release.